requirements.txt
asgiref==3.8.1
attrs==25.4.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
dj-database-url==3.0.1
Django==5.2.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.9.1
geographiclib==2.1
geopy==2.4.1
google-auth==2.40.2
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Markdown==3.8
packaging==25.0
pillow==12.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
pyproj==3.7.1
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.37.0
requests==2.32.3
rpds-py==0.30.0
rsa==4.9.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
whitenoise==6.10.0
